services/api-gateway/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.32.5
google-generativeai==0.8.5









